feat: Enable multiselect search results for tissue (#3234)
